Extract security tests into own file
